+++ src/commands/mapshaper-blur.mjs
@@ -1,5 +1,5 @@
 import cmd from '../mapshaper-cmd';
-import { requireProjectedDataset } from '../crs/mapshaper-projections';
+import { getDatasetCRS, requireProjectedDataset } from '../crs/mapshaper-projections';
 import { layerHasRaster } from '../dataset/mapshaper-layer-utils';
 import { blurRasterGrid } from '../rasters/mapshaper-raster-blur';
 import { createRasterPreview } from '../rasters/mapshaper-raster-utils';
@@ -14,19 +14,20 @@ cmd.blur = blurRasterLayers;
 
 export function blurRasterLayers(layers, dataset, optsArg) {
   var opts = optsArg || {};
+  var crs = getDatasetCRS(dataset);
   requireProjectedDataset(dataset);
   layers.forEach(function(lyr) {
     if (!layerHasRaster(lyr)) {
       stop('Command requires a raster layer');
     }
-    blurRasterLayer(lyr, opts);
+    blurRasterLayer(lyr, opts, crs);
   });
 }
 
-function blurRasterLayer(lyr, opts) {
+function blurRasterLayer(lyr, opts, crs) {
   var raster = lyr.raster;
   noteLayerWillChange(lyr, {operation: 'blurRasterLayer', unit: 'raster'});
-  raster.grid = blurRasterGrid(raster, opts);
+  raster.grid = blurRasterGrid(raster, opts, crs);
   raster.view = raster.view || {};
   delete raster.view.scalingStats;
   if (runningInBrowser()) {