+++ src/mapshaper-internal.mjs
@@ -129,6 +129,7 @@ import * as LayerUtils from './dataset/mapshaper-layer-utils';
 import * as Lines from './commands/mapshaper-lines';
 import * as Logging from './utils/mapshaper-logging';
 import * as Profile from './utils/mapshaper-profile';
+import * as RasterBlur from './rasters/mapshaper-raster-blur';
 import * as RasterContours from './rasters/mapshaper-raster-contours';
 import * as RasterGrid from './rasters/mapshaper-raster-grid';
 import * as RasterReprojection from './rasters/mapshaper-raster-reprojection';
@@ -263,6 +264,7 @@ Object.assign(internal,
   Lines,
   Logging,
   Profile,
+  RasterBlur,
   RasterContours,
   RasterGrid,
   RasterReprojection,